change datatype in validator from int to float32

+++ weather_service/internal/validator/validator.go
@@ -1,22 +1,22 @@
 package validator
 
 type Validator struct {
-	LatitudeMin  int
-	LatitudeMax  int
-	LongitudeMin int
-	LongitudeMax int
+	LatitudeMin  float32
+	LatitudeMax  float32
+	LongitudeMin float32
+	LongitudeMax float32
 }
 
 func New() Validator {
 	return Validator{
-		LatitudeMin:  -90,
-		LatitudeMax:  90,
-		LongitudeMin: -180,
-		LongitudeMax: 180,
+		LatitudeMin:  -90.0,
+		LatitudeMax:  90.0,
+		LongitudeMin: -180.0,
+		LongitudeMax: 180.0,
 	}
 }
 
-func (v *Validator) ValidateCoordinates(latitude, longitude int) error {
+func (v *Validator) ValidateCoordinates(latitude, longitude float32) error {
 	if v.LatitudeMin > latitude || v.LatitudeMax < latitude {
 		return ErrIncorrectLat
 	}
